+++ Runtime/UI/Scripts/Vector2UI.cs
@@ -34,6 +34,8 @@ public class Vector2UI : ControllableUI
 
         XInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getFloat(value.ToString()));
             list.Add(TypeConverter.getFloat(YInput.text.ToString()));
@@ -43,6 +45,8 @@ public class Vector2UI : ControllableUI
 
         YInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getFloat(XInput.text.ToString()));
             list.Add(TypeConverter.getFloat(value.ToString()));