+++ Runtime/UI/Scripts/Vector3IntUI.cs
@@ -32,6 +32,8 @@ public class Vector3IntUI : ControllableUI
 
         XInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getInt(value.ToString()));
             list.Add(TypeConverter.getInt(YInput.text.ToString()));
@@ -42,6 +44,8 @@ public class Vector3IntUI : ControllableUI
 
         YInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getInt(XInput.text.ToString()));
             list.Add(TypeConverter.getInt(value.ToString()));
@@ -52,6 +56,8 @@ public class Vector3IntUI : ControllableUI
 
         ZInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getInt(XInput.text.ToString()));
             list.Add(TypeConverter.getInt(YInput.text.ToString()));