+++ Runtime/UI/Scripts/Vector3UI.cs
@@ -40,6 +40,8 @@ public class Vector3UI : ControllableUI
 
         XInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getFloat(value.ToString()));
             list.Add(TypeConverter.getFloat(YInput.text.ToString()));
@@ -50,6 +52,8 @@ public class Vector3UI : ControllableUI
 
         YInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getFloat(XInput.text.ToString()));
             list.Add(TypeConverter.getFloat(value.ToString()));
@@ -60,6 +64,8 @@ public class Vector3UI : ControllableUI
 
         ZInput.onEndEdit.AddListener((value) =>
         {
+            RecordUndo();
+
             var list = new List<object>();
             list.Add(TypeConverter.getFloat(XInput.text.ToString()));
             list.Add(TypeConverter.getFloat(YInput.text.ToString()));